modularizing code part3
